Get BirdScan visibility table

Load visibility table from an already connected 'Birdscan MR1' 'SQL' database.

Usage

getVisibilityTable(dbConnection, dbDriverChar)

Arguments

dbConnection

A valid database connection.

dbDriverChar

This was the name of the driver, and is now automatically detected, therefore it should be omitted into the future.

Value

A data.frame with the visibility table.

See also

Author

Fabian Hertner, Birgen Haest

Examples

if (FALSE) { # \dontrun{
# Set server and database settings
# ===========================================================================
dbServer = "MACHINE\\SERVERNAME" # Set the name of your SQL server
dbName = "db_Name" # Set the name of your database
dbDriverChar = "SQL Server" # Set either "SQL Server" or "PostgreSQL"

# Open the connection with the database
# ===========================================================================
dsn = paste0(
  "driver=", dbDriverChar, ";server=", dbServer,
  ";database=", dbName,
  ";uid=", rstudioapi::askForPassword("Database user"),
  ";pwd=", rstudioapi::askForPassword("Database password")
)
dbConnection = RODBC::odbcDriverConnect(dsn)

visibilityTable = getVisibilityTable(dbConnection)
} # }